Make swing global
Before this commit, when we applied swing we directly modified the
`offset` of all even steps. This commit changes it so that swing is a
parameter that is stored on the Pattern and each of its tracks. This
means that if we happen to create new steps (currently we pre-populate
all 64 steps in a track), they will still have swing correctly
applied, and we can also vary the swing of individual tracks.