Clamp editable step parameters to sensible range
We don't want to be able to offset a step more than the number of
ticks in a 16th note (I'm ignoring the possibility that notes might
already be offset due to swing for now).

I've set the velocity range 0 -> 127 to match midi, and then scaled it
between 0 and 1 for timber's input.