Move amplitude envelope into its own synthdef
I want to have different sample playback engines for looping
samples. This change makes it easier to do that by moving the
ampltiude envelope out of the playback synth.

I've switched from a ASR envelope to a percussive one because the
latter doesn't rely on knowing how long the sustain portion of the
envelope needs to be. In the case of looping samples that I want to
introduce we will need to specify the loop duration anyway, so this
change will work for now. I've scaled the 0-1 input range for the
release potion to 5s max.