Create sample playback synths on the fly
Now we create a "oneshot" synth each time a note on message is
received and set the sample start and stop on the player when we
create it. That should make it easier to eventually introduce a
looping synth associated with the same track.

For now, the synths are freed when they have finished playing their
buffer, which means that samples on the same track can currently
overlap. We'll need to introduce something to free them explicitly
when new ones are triggered before we can have looping synths.