Free any playing synths when a new note is triggered
This ensures that any notes triggered stop currently playing notes for
that track. This effectively limits the polyphony of the engine to the
number of tracks (which is how the digitakt / model:samples work as
far as I know). It also makes it easier to introduce a looping sample
player, as new triggers can then stop the loop.