Change amplitude envelope to have a duration
For now I think this is a sensible envelope shape to have. I've set
the duration to be 0.25s (which is 1/2 beat at 120bpm), with the idea
that we could set a duration per-track in terms of the number of steps
at the current tempo.

I think a more flexible approach (and one that I might have to do
later) is to have a `note_off` message that causes the envelopes to
enter the release phase. But that would involve some reworking of the
sequencer to send note off messages for each note on at the right
time. I think I'll leave that for now.