Release players rather than free them
When a new player is triggered on the same track as an existing one,
we now "release" it rather than "free" it. This is possible because
both oneshotplayer and loopplayer now have a gate argument, which is
required by release.

Making this change removes the clicking that can occur when a playing
buffer is immediately freed. But it also means that new triggered
players overlap the previous ones. As long as we don't hit any
performance limitations (too many players) I think this is fine. It's
up to the user of thunk to set the duration and envelope parameters to
use this in a musical way.